+++ app/controllers/article-category.js
@@ -1,6 +1,7 @@
 'use strict'
 const Services = require('../services')
-const {auth, resHandler} = require('../myutil')
+const {auth, resHandler, paramsHandler} = require('../myutil')
+const {pageConfig} = require('../../config')
 
 class ArticleCategoryController {
   async create (req, res) {
@@ -8,6 +9,35 @@ class ArticleCategoryController {
       const userInfo = auth.verifyToken(req.headers.token)
       req.body.userId = userInfo.userId
       const result = await Services.articleCategory.addCategory(req.body)
+      res.sendOk(result)
+    } catch (error) {
+      const errorRes = resHandler.getErrorRes(error)
+      res.sendErr(errorRes)
+    }
+  }
+
+  async update (req, res) {
+    try {
+      const result = await Services.articleCategory.editById(req.params._id, req.body)
+      res.sendOk(result)
+    } catch (error) {
+      const errorRes = resHandler.getErrorRes(error)
+      res.sendErr(errorRes)
+    }
+  }
+
+  async list (req, res) {
+    try {
+      // 翻页参数处理
+      const offset = paramsHandler.offsetFormat(req.query, pageConfig.users)
+      console.log(req.query)
+      const queryObj = {
+        condition: req.query,
+        skipCount: offset.skipCount,
+        pagesize: offset.pagesize,
+        sortRule: offset.sortRule
+      }
+      const result = await Services.articleCategory.getCategoryList(queryObj)
       res.sendOk(result)
     } catch (error) {
       const errorRes = resHandler.getErrorRes(error)