+++ app/services/index.js
@@ -2,7 +2,6 @@
 const Services = {}
 
 Services.users = require('./users')
-Services.db = require('./db')
 Services.articleCategory = require('./article-category')
 
 module.exports = Services