+++ config/pagesize.js
@@ -1,5 +1,6 @@
 'use strict'
 
 module.exports = {
-  users: 10
+  users: 10,
+  category: 10
 }