+++ config/success-message.js
@@ -1,8 +1,9 @@
 'use strict'
 module.exports = {
-  USER_UPDATE_SUCCESS: '用户信息更新成功',
-  USER_DELETE_SUCCESS: '用户删除成功',
-  OPTION_SUCCESS: '操作成功',
-  ROUTER_DELETE_SUCCESS: '路由删除成功',
-  ROUTER_UPDATE_SUCCESS: '路由更新成功'
+  USER_UPDATE_SUCCESS: '用户信息更新成功！',
+  USER_DELETE_SUCCESS: '用户删除成功！',
+  OPTION_SUCCESS: '操作成功！',
+  ROUTER_DELETE_SUCCESS: '路由删除成功！',
+  ROUTER_UPDATE_SUCCESS: '路由更新成功！',
+  CATEGORY_UPDATE_SUCCESS: '文章类别修改成功！'
 }